% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pedigree.R
\name{simulate_pedigree}
\alias{simulate_pedigree}
\title{simulate_pedigree}
\usage{
simulate_pedigree(
  years = 5,
  n_females = 50,
  afr = 1,
  p_breed = 1,
  fecundity = 4,
  fixed_fecundity = TRUE,
  p_polyandry = 0,
  p_sire = 1,
  p_retain = 0,
  v_EPP = 0,
  juv_surv = 0.25,
  adult_surv = 0.5,
  immigration = 0,
  constant_pop = TRUE,
  known_age_structure = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{years}{Integer. Number of time steps/synchronous reproductive events}

\item{n_females}{Integer. Starting number of breeding females}

\item{afr}{Integer. Age at first reproduction}

\item{p_breed}{Probability that a female breeds, must be between 0 and 1. Can be sex-specific if entered as a vector of length 2 (female,male).}

\item{fecundity}{Mean number of juveniles a female produces each year.}

\item{fixed_fecundity}{Logical. is fecundity fixed or drawn from a Poisson distribution}

\item{p_polyandry}{Probability that a female has any polyandry. Must be between 0 and 1.}

\item{p_sire}{Probability that 'social' male sires all offspring. Must be between 0 and 1.}

\item{p_retain}{Probability that social partnership is retained. Must be between 0 and 1.}

\item{juv_surv}{survival of juveniles until local recruitment, where recruitment is defined as having genetic offspring. Must be between 0 and 1. Can be sex-specific if entered as a vector of length 2 (female,male).}

\item{adult_surv}{survival of adults across years. Must be between 0 and 1. Can be sex-specific if entered as a vector of length 2 (female,male).}

\item{immigration}{Yearly immigration, as a proportion of starting number of females (n_females). Must be between 0 and 1. Can be sex-specific if entered as a vector of length 2 (female,male).}

\item{constant_pop}{Logical. Should there be stochastic variation in population size}

\item{known_age_structure}{Currently not in use}

\item{verbose}{Logical - print simulation progress? useful for debugging}
}
\value{
A list with two elements: population data and a pedigree
}
\description{
Individual based simulation based on specified demographic parameters.
}
\details{
...
}
\examples{
\dontrun{
}
}
\author{
Joel Pick - joel.l.pick@gmail.com
}
